Do not report an error when probing fails if it will be retried
Prior to this commit, whenever the first attempt to probe (tested
on a Taz 6) failed, an error was written to the serial channel
and this error causes Octoprint to abort the print job.

A workaround would be to tell Octoprint to ignore "Probing failed"
error messages, that seems less than ideal.

This commit removes the standard Marlin error reporting since
the Lulzbot reprobing code will generate an error as needed.

This allows a rewiping to succeed when printing via Octoprint
but still allows an error to be reporting if probing truly fails.